# docs/OPENAPI_SWAGGER_SPECIFICATION.yml
openapi: 3.0.3
info:
  title: CI/CD Status Notifier Bot API
  description: |
    Comprehensive REST API for the CI/CD Status Notifier Bot system. This API provides endpoints for managing CI/CD projects, processing webhooks, handling Telegram bot interactions, and monitoring system metrics.
    
    ## Features
    - **Project Management**: Complete CRUD operations for CI/CD projects
    - **Webhook Processing**: GitHub webhook event handling and processing
    - **Telegram Bot Integration**: Bot commands and notification management
    - **Dashboard Analytics**: Real-time metrics and build analytics
    - **Health Monitoring**: System health and status endpoints
    
    ## Architecture
    - **Backend**: Go with Fiber framework
    - **Database**: PostgreSQL with GORM
    - **Architecture Pattern**: Hexagonal/Clean Architecture
    - **Authentication**: Webhook signature verification (HMAC-SHA256)
  version: "1.0.0"
  contact:
    name: CICD Status Notifier Bot Team
    email: support@cicd-notifier.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.cicd-notifier.com
    description: Production server
paths:
...
  /api/v1/projects:
    post:
      tags:
        - Project Management
      summary: Create a new project
      description: Creates a new CI/CD project with the provided information
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
            examples:
              basic_project:
                summary: Basic project creation
                value:
                  name: "My Test Project"
                  repository_url: "https://github.com/user/test-repo"
                  webhook_secret: "my-secure-secret-key"
              with_telegram:
                summary: Project with Telegram integration
                value:
                  name: "Production API"
                  repository_url: "https://github.com/company/production-api"
                  webhook_secret: "super-secure-webhook-secret"
                  telegram_chat_id: -1001234567890
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Project created successfully"
                  data:
                    $ref: '#/components/schemas/ProjectResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Project Management
      summary: List projects
      description: Retrieves a list of projects with optional filtering and pagination
      operationId: listProjects
      parameters:
        - name: status
          in: query
          description: Filter by project status
          schema:
            type: string
            enum: [active, inactive, archived]
        - name: name
          in: query
          description: Filter by project name (partial match)
          schema:
            type: string
        - name: repository_url
          in: query
          description: Filter by repository URL (partial match)
          schema:
            type: string
        - name: has_telegram_chat
          in: query
          description: Filter by presence of Telegram chat integration
          schema:
            type: boolean
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: offset
          in: query
          description: Number of results to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sort_by
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [name, repository_url, status, created_at, updated_at]
            default: created_at
        - name: sort_order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Projects retrieved successfully"
                  data:
                    $ref: '#/components/schemas/ListProjectResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/projects/{id}/statistics:
    get:
      tags:
        - Dashboard Analytics
      summary: Get project statistics
      description: Retrieves detailed build statistics and metrics for a specific project
      operationId: getProjectStatistics
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Project statistics retrieved successfully"
                  data:
                    $ref: '#/components/schemas/ProjectStatisticsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Webhook Endpoints
...
  /api/v1/dashboard/overview:
    get:
      tags:
        - Dashboard Analytics
      summary: Get dashboard overview
      description: Retrieves comprehensive overview metrics for the entire CI/CD system
      operationId: getDashboardOverview
      responses:
        '200':
          description: Dashboard overview retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Dashboard overview retrieved successfully"
                  data:
                    $ref: '#/components/schemas/OverviewResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/builds/analytics:
    get:
      tags:
        - Dashboard Analytics
      summary: Get build analytics
      description: Retrieves comprehensive build analytics and trends over a specified time range
      operationId: getBuildAnalytics
      parameters:
        - name: range
          in: query
          description: Time range for analytics
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 7d
      responses:
        '200':
          description: Build analytics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Build analytics retrieved successfully"
                  data:
                    $ref: '#/components/schemas/BuildAnalyticsResponse'
        '400':
          description: Invalid time range
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid time range"
                  details:
                    type: string
                    example: "Supported ranges are: 7d, 30d, 90d"
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Telegram Bot Endpoints
...
components:
  schemas:
...
    CreateProjectRequest:
      type: object
      required:
        - name
        - repository_url
        - webhook_secret
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Unique name for the project
          example: "My Test Project"
        repository_url:
          type: string
          format: uri
          description: GitHub repository URL
          example: "https://github.com/user/test-repo"
        webhook_secret:
          type: string
          minLength: 10
          description: Secret key for webhook signature verification
          example: "my-secure-secret-key"
        telegram_chat_id:
          type: integer
          format: int64
          description: Optional Telegram chat ID for notifications
          example: -1001234567890
...
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique project identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          description: Project name
          example: "My Test Project"
        repository_url:
          type: string
          format: uri
          description: GitHub repository URL
          example: "https://github.com/user/test-repo"
        status:
          type: string
          enum: [active, inactive, archived]
          description: Current project status
          example: "active"
        telegram_chat_id:
          type: integer
          format: int64
          nullable: true
          description: Telegram chat ID for notifications
          example: -1001234567890
        created_at:
          type: string
          format: date-time
          description: Project creation timestamp
          example: "2025-07-29T09:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-07-29T09:00:00Z"
    ListProjectResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
        total:
          type: integer
          format: int64
          description: Total number of projects matching filters
          example: 25
        limit:
          type: integer
          description: Number of projects per page
          example: 10
        offset:
          type: integer
          description: Number of projects skipped
          example: 0
...
    OverviewResponse:
      type: object
      properties:
        total_projects:
          type: integer
          description: Total number of projects
          example: 50
        active_projects:
          type: integer
          description: Number of active projects
          example: 42
        total_builds:
          type: integer
          description: Total number of builds
          example: 1250
        successful_builds:
          type: integer
          description: Number of successful builds
          example: 1125
        failed_builds:
          type: integer
          description: Number of failed builds
          example: 125
        success_rate:
          type: number
          format: float
          description: Build success rate (0-100)
          example: 90.0
        average_duration:
          type: integer
          description: Average build duration in seconds
          example: 180
        last_updated:
          type: string
          format: date-time
          description: Last metrics update timestamp
          example: "2025-08-01T10:00:00Z"
    ProjectStatisticsResponse:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        project_name:
          type: string
          example: "My Test Project"
        total_builds:
          type: integer
          example: 150
        successful_builds:
          type: integer
          example: 135
        failed_builds:
          type: integer
          example: 15
        success_rate:
          type: number
          format: float
          example: 90.0
        average_duration:
          type: integer
          description: Average build duration in seconds
          example: 180
        last_build_time:
          type: string
          format: date-time
          example: "2025-08-01T09:45:00Z"
        build_trends:
          type: array
          items:
            $ref: '#/components/schemas/BuildTrendData'
    BuildTrendData:
      type: object
      properties:
        date:
          type: string
          format: date
          example: "2025-08-01"
        count:
          type: integer
          description: Total builds on this date
          example: 25
        success_count:
          type: integer
          description: Successful builds on this date
          example: 22
    BuildAnalyticsResponse:
      type: object
      properties:
        time_range:
          type: string
          description: Analytics time range
          example: "7d"
        total_builds:
          type: integer
          example: 500
        successful_builds:
          type: integer
          example: 450
        failed_builds:
          type: integer
          example: 50
        success_rate:
          type: number
          format: float
          example: 90.0
        average_duration:
          type: integer
          description: Average build duration in seconds
          example: 180
        builds_by_day:
          type: array
          items:
            $ref: '#/components/schemas/BuildAnalyticsData'
        duration_trends:
          type: array
          items:
            $ref: '#/components/schemas/DurationTrendData'
    BuildAnalyticsData:
      type: object
      properties:
        date:
          type: string
          format: date
          example: "2025-08-01"
        total_builds:
          type: integer
          example: 75
        successful_builds:
          type: integer
          example: 68
        failed_builds:
          type: integer
          example: 7
    DurationTrendData:
      type: object
      properties:
        date:
          type: string
          format: date
          example: "2025-08-01"
        average_duration:
          type: integer
          description: Average build duration in seconds
          example: 185

    # Telegram Schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Validation failed"
        details:
          type: string
          description: Detailed error information
          example: "Field 'name' is required"
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      description: Unique project identifier (UUID)
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              summary: Validation error example
              value:
                error: "Validation failed"
                details: "Field 'name' is required"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            not_found:
              summary: Not found error example
              value:
                error: "project not found"
    ConflictError:
      description: Resource already exists or conflict occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            conflict:
              summary: Conflict error example
              value:
                error: "project with this name already exists"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server_error:
              summary: Internal server error example
              value:
                error: "Internal server error"
  securitySchemes:
    GitHubWebhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: |
        GitHub webhook signature for request verification. The signature is calculated using HMAC-SHA256 with the webhook secret.
        
        Format: `sha256={signature}`
        
        Example: `sha256=abc123def456...`
...